SLVUU-49: Extract LayoutTile
